% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabrowska.R
\name{dabrowska}
\alias{dabrowska}
\title{Multivariate Dabrowska estimator}
\usage{
dabrowska(data, k, ...)
}
\arguments{
\item{data}{An n x 2k data frame with the following items:
\describe{
\item{X1, X2, ..., Xk:}{Survival times for k dimensions.}
\item{Delta1, Delta2, ..., Deltak:}{Censoring indicators for k dimensions.}
}}

\item{k}{The number of dimensions.}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
A data frame containing univariate, bivariate, and trivariate survival statistics.
}
\description{
Calculates the k-variate survival function.
}
\examples{

data <- genClaytonk(10, 1, c(0.1, 0.2, 0.3))
trivariate <- dabrowska(data, 3)

}
