% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAME-package.R
\docType{package}
\name{DAME-package}
\alias{DAME-package}
\title{DAME-package:  Marginal Effects in Nonlinear Interactive Models.}
\description{
The DAME package provides tools that implement the procedures for computing and visualizing marginal effects of independent variables
in nonlinear GLMs, as described in Zhirnov, Moral, Sedashov (2021).
}
\references{
Zhirnov, Andrei, Mert Moral, and Evgeny Sedashov (2021). ``Taking Distributions Seriously: On the Interpretation
of the Estimates of Interactive Nonlinear Models.'' Working paper.
}
\author{
{
\strong{Maintainer}: Evgeny Sedashov \email{esedashov@hse.ru}
\itemize{
\item Mert Moral \email{mmoral@sabanciuniv.edu}
\item Andrei Zhirnov \email{A.Zhirnov@exeter.ac.uk}
}
}
}
