% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genClaytonk.R
\name{genClaytonk}
\alias{genClaytonk}
\title{Multivariate survival data simulation from Clayton-Oakes model}
\usage{
genClaytonk(n, theta, lambdaC)
}
\arguments{
\item{n}{Number of observations to generate.}

\item{theta}{Clayton copula parameter (-1 <= theta < inf).}

\item{lambdaC}{Vector of exponentially-distributed rate parameters for censoring times.
(0 = censored, 1 = uncensored)}
}
\value{
An n x 2k data frame with the following items:
\describe{
\item{X1, X2, ..., Xk:}{Survival times for k dimensions.}
\item{Delta1, Delta2, ..., Deltak:}{Censoring indicators for k dimensions.}
}
}
\description{
Generates simulated k-variate survival data from a Clayton-Oakes model.
The marginals are exponentially distributed with rate parameters 1. 
The joint distribution follows from the Clayton copula.
Censoring times are exponentially distributed with vector of rate parameters lambdaC.
}
\examples{
X <- genClaytonk(200, 3, c(0.1, 0.2, 0.3, 0.4))

}
