% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core2.R
\name{GPpost}
\alias{GPpost}
\title{GPpost}
\usage{
GPpost(kappa, parameter, tau, dist_ij, a, b)
}
\arguments{
\item{kappa}{current value of kappa}

\item{parameter}{parameter which kappa is embedded}

\item{tau}{scale parameter of Gaussian process covariance}

\item{dist_ij}{matrix of squared distance between all timepoints}

\item{a}{inverse-gamma shape parameter}

\item{b}{inverse-gamma scale parameter}
}
\value{
log- posterior density of kappa
}
\description{
log-posterior distribution of kappa in case of multiple dimensions
}
