% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core2.R
\name{rd_fc2}
\alias{rd_fc2}
\title{rd_fc2}
\usage{
rd_fc2(XB, theta, U, d, Y, cinv, s2, tau_r, meaningful_NA)
}
\arguments{
\item{XB}{Time-length list of XB matrix}

\item{theta}{a T x n matrix of nodal effects}

\item{U}{current value of U}

\item{d}{current value of d}

\item{Y}{relational array of relations}

\item{cinv}{inverse of Gaussian process covariance matrix}

\item{s2}{variance of normal error}

\item{tau_r}{scale parameter of Gaussian process covariance}

\item{meaningful_NA}{position of structural NA}
}
\value{
One sample estimate of d from normal distribution
}
\description{
Sample d with structural NA
}
