% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core2.R
\name{rtaui_fc}
\alias{rtaui_fc}
\title{rtaui_fc}
\usage{
rtaui_fc(theta, cinv, a, b)
}
\arguments{
\item{theta}{a Time by n vector of nodal effects}

\item{cinv}{inverse of Gaussian process covariance matrix}

\item{a}{inverse-gamma shape parameter}

\item{b}{inverse-gamma scale parameter}
}
\value{
One sample estimate of tau_i from inverse-gamma distribution
}
\description{
Sample the variance in theta
}
