% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core2.R
\name{rtaup_fc}
\alias{rtaup_fc}
\title{rtaup_fc}
\usage{
rtaup_fc(beta, cinv, a, b)
}
\arguments{
\item{beta}{a p by 1 vector of coefficients}

\item{cinv}{inverse of Gaussian process covariance matrix}

\item{a}{inverse-gamma shape parameter}

\item{b}{inverse-gamma scale parameter}
}
\value{
One sample estimate of tau_p from inverse-gamma distribution
}
\description{
Sample the variance in beta
}
