% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core2.R
\name{ru_fc2}
\alias{ru_fc2}
\title{ru_fc2}
\usage{
ru_fc2(XB, theta, U, Y, tau_u, s2, meaningful_NA_rows)
}
\arguments{
\item{XB}{Time-length list of XB matrix}

\item{theta}{a T x n matrix of nodal effects}

\item{U}{current value of U}

\item{Y}{relational array of relations}

\item{tau_u}{scale parameter of Gaussian process covariance}

\item{s2}{variance of normal error}

\item{meaningful_NA_rows}{position of structural NA}
}
\value{
One sample estimate of d from normal distribution
}
\description{
Sample u with structural NA
}
