% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{strikes}
\alias{strikes}
\title{This is an adapted and simplified version of the dataset used in Robertson, G. B., and E. Teitelbaum. 2011.
"Foreign Direct Investment, Regime Type, and Labor Protest in Developing Countries." American Journal of Political Science 55 (3): 665–677.}
\format{
A data frame with 1507 rows and 10 variables:
\describe{
  \item{isocode}{Numeric ISO code for the country}
  \item{iso3c}{3-letter ISO code for the country}
  \item{year}{Year}
  \item{flows.ln}{Natural log of FDI flows (lagged)}
  \item{population.ln}{Natural log of the population}
  \item{polity2}{Polity score}
  \item{disputes}{Number of industrial strikes}
  \item{any_dispute}{Any binary strike in a year}
  \item{gdp_grth}{GDP growth}
  \item{gdp_pc}{GDP per capita}
}
}
\source{
Robertson, G. B., and E. Teitelbaum. 2011.
"Foreign Direct Investment, Regime Type, and Labor Protest in Developing Countries." American Journal of Political Science 55 (3): 665–677.
}
\usage{
strikes
}
\description{
This is an adapted and simplified version of the dataset used in Robertson, G. B., and E. Teitelbaum. 2011.
"Foreign Direct Investment, Regime Type, and Labor Protest in Developing Countries." American Journal of Political Science 55 (3): 665–677.
}
\keyword{datasets}
