% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAME-package.R
\docType{package}
\name{DAME-package}
\alias{DAME-package}
\title{DAME-package:  Marginal Effects in Nonlinear Interactive Models.}
\description{
The DAME package provides tools that implement and visualize procedures described in Moral, Sedashov
and Zhirnov (2017). It can also be used for computing other well-known quantities of interest such
as predicted probabilities.
}
\references{
Moral, Mert, Evgeny Sedashov, and Andrei Zhirnov. (2017) ``Taking Distributions Seriously
Interpreting the Effects of Constitutive Variables in Nonlinear Models with Interactions.'' Working paper.
}
\author{
{
\strong{Maintainer}: Evgeny Sedashov \email{esedash1@binghamton.edu}

Authors:
\itemize{
\item Mert Moral \email{mmoral17@binghamton.edu}
\item Andrei Zhirnov \email{azhirno1@binghamton.edu}
}
}
}
