% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dame.R
\name{dame}
\alias{dame}
\title{DAME function}
\usage{
dame(
  x,
  over = NULL,
  model = NULL,
  data = NULL,
  formula = NULL,
  link = NULL,
  coefficients = NULL,
  variance = NULL,
  nbins = 10,
  bin_id = NULL,
  use_distinct_values = FALSE,
  discrete = FALSE,
  discrete_step = 1,
  at = NULL,
  mc = FALSE,
  pct = c(2.5, 97.5),
  iter = 1000
)
}
\arguments{
\item{x}{A character string representing the name of a variable. DAME will be computed for this variable.}

\item{over}{A character string representing the name of a variable. DAME will be computed for each of the values of this variable.}

\item{model}{Fitted model object. The package works best with GLM objects and will extract the formula, dataset, family, coefficients, and
the QR components of the design matrix if arguments \code{formula}, \code{data}, \code{link}, \code{coefficients}, and/or
\code{variance} are not explicitly specified.}

\item{data}{The dataset to be used to compute marginal effects. DAME provides most meaningful results if the original dataset is used for computing marginal effects.}

\item{formula}{The formula used in estimation (if not specified, it is extracted from the fitted model object).}

\item{link}{The link function used in estimation (if not specified, it is extracted from the fitted model object).}

\item{coefficients}{The named vector of coefficients produced during the estimation (if not specified, it is extracted from the fitted model object).}

\item{variance}{The variance-covariance matrix to be used for computing standard errors (if not specified, it is extracted from the fitted model object).}

\item{nbins}{the number of bins to be used for aggregating marginal effects; the default is 10 bins of equal size; ignored if \code{bin_id} is specified or
\code{use_distinct_values} is TRUE.}

\item{bin_id}{a numeric vector identifying the bins used for aggregating marginal effects.}

\item{use_distinct_values}{logical; if TRUE, the function uses all unique values of the \code{over} variable; ignored if \code{bin_id} is specified.}

\item{discrete}{A logical variable. If TRUE, the function will compute the effect of a discrete change in \code{x}. If FALSE, the function will compute the partial derivative of \code{x}.}

\item{discrete_step}{The size of a discrete change in \code{x} used in computations (used only if \code{discrete=TRUE}).}

\item{at}{A named list of values of independent variables. These variables will be set to these value before computations. All other quantitative variables (except \code{x} and \code{over}) will be set to their means. All other factor variables will be set to their modes.}

\item{mc}{A logical variable. If TRUE, the function will compute standard errors and sampling quantiles using Monte-Carlo simulations. If FALSE, the function will use the delta method.}

\item{pct}{A numeric vector with the sampling quantiles to be output with the DAME estimates. Default = \code{c(2.5,97.5)}.}

\item{iter}{Number of interations used in Monte-Carlo simulations. Default = 1,000.}
}
\value{
A list of the following:
\itemize{
\item\code{dame} A data frame with the DAME estimates, standard errors, quantiles of the sampling distribution, and the values of the independent variables used for computing DAME.
\item\code{execute_time} Execution time
}
}
\description{
\code{dame} computes the distribution-weighted average marginal effects (DAME) as described in Moral, Sedashov, and Zhirnov (2017).
}
\examples{
##poisson regression with 2 variables and an interaction between them
#fit the regression first
data <- data.frame(y = rpois(10000, 10), x2 = rpois(10000, 5), x1 = rpois(10000, 3))
y <- glm(y ~ x1*x2, data = data, family = "poisson")
#compute DAME
dame(model = y, x = "x1", over = "x2")
}
\references{
Greene, William. 2012. \emph{Econometric Analysis, 7 ed.} Pearson Education Limited.

King, Garry, Michael Tomz, and Jason Wittenberg. 2000. ``Making the Most of Statistical Analyses: Improving Interpretation and Presentation.'' \emph{American Journal of Political Science} 44(2): 341-355.
Moral, Mert, Evgeny Sedashov, and Andrei Zhirnov. (2017) ``Taking Distributions Seriously Interpreting the Effects of Constitutive Variables in Nonlinear Models with Interactions.'' Working paper.
}
\author{
Function \code{dame} comes as a complimentary open-source implementation of procedures described in Moral, Sedashov, and Zhirnov (2017).
Standard errors are computed using either delta method (Greene 2012) for more details) or Monte-Carlo simulations (King, Tomz, and Wittenberg 2000).
}
