% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_me.R
\name{plot_me}
\alias{plot_me}
\title{Vizualizing Marginal Effects}
\usage{
plot_me(
  x,
  over,
  model = NULL,
  data = NULL,
  link = NULL,
  formula = NULL,
  coefficients = NULL,
  variance = NULL,
  at = NULL,
  mc = FALSE,
  iter = 1000,
  heatmap_dim = c(100, 100),
  p = 0.05,
  gradient = c("#f2e6e7", "#f70429"),
  breaks_hor = waiver(),
  breaks_ver = waiver(),
  breaks_hb = waiver(),
  breaks_hl = waiver(),
  geom_hl = geom_histogram(aes(y = stat(count)/sum(stat(count)))),
  geom_hb = geom_histogram(aes(y = stat(count)/sum(stat(count)))),
  theme_mp = theme_bw() + theme(panel.grid.minor = element_blank(), panel.grid.major =
    element_blank(), panel.border = element_rect(colour = "black"), axis.text.y =
    element_text(angle = 90), axis.title.x = element_blank(), axis.title.y =
    element_blank(), plot.margin = unit(c(0.5, 0.5, 0.01, 0.01), "lines"), aspect.ratio =
    1, legend.position = "right"),
  theme_hl = theme_bw() + theme(panel.border = element_rect(colour = NA),
    panel.grid.minor = element_blank(), panel.grid.major = element_blank(),
    axis.ticks.length = unit(0.1, "cm"), plot.margin = unit(c(0.5, 0, 0, 0.5), "lines"),
    axis.title.y = element_text(face = "bold"), axis.title.x = element_blank(),
    aspect.ratio = 5),
  theme_hb = theme_bw() + theme(panel.border = element_rect(colour = NA),
    panel.grid.minor = element_blank(), panel.grid.major = element_blank(),
    axis.ticks.length = unit(0.1, "cm"), plot.margin = unit(c(0, 0.5, 0, 0), "lines"),
    axis.text.y = element_text(angle = 90), axis.title.x = element_text(face = "bold"),
    axis.title.y = element_blank(), aspect.ratio = 1/5)
)
}
\arguments{
\item{x}{The independent variable whose effect is being assessed. This variable is mapped along the vertical axis.}

\item{over}{The second independent variable. This variable is mapped along the horizontal axis.}

\item{model}{Fitted model object. The package works best with GLM objects and will extract the formula, dataset, family, coefficients, and
the QR components of the design matrix if arguments \code{formula}, \code{data}, \code{link}, \code{coefficients}, and/or
\code{variance} are not explicitly specified.}

\item{data}{The dataframe used in estimation.}

\item{at}{A named list of the values for other indepedent variables. Unless listed, other continuous variables
are fixed at their means, and all factor variables are set to their modes.}

\item{mc}{A logical variable. If TRUE, the function will compute standard errors and sampling quantiles using Monte-Carlo simulations. If FALSE, the function will use the delta method.}

\item{iter}{Number of interations used in Monte-Carlo simulations. Default = 1,000.}

\item{heatmap_dim}{Numeric vector containin the number of rows and columns used for drawing the heatmap. Default = 100 each.}

\item{p}{Singificance level for the marginal effects. Default = 0.05.}

\item{gradient}{String vector representing color gradient for the contour plot. First element
is the color of the lowest value, and second element is the color of the highest value.}

\item{breaks_hor}{Numeric vector of breaks for the horizontal axis ("over" variable) \code{waiver()} by default.}

\item{breaks_ver}{Numeric vector of breaks for the vertical axis ("x" variable)  \code{waiver()} by default.}

\item{breaks_hb}{Numeric vector specifying probability axis for the bottom histogram.
\code{waiver()} by default.}

\item{breaks_hl}{Numeric vector specifying probability axis for the left histogram.
\code{waiver()} by default.}

\item{geom_hl}{Specifies ggplot layer of the type \code{geom_histogram} for the left histogram.
If NULL, default settings are used.}

\item{geom_hb}{Specifies ggplot layer of the type \code{geom_histogram} for the bottom histogram.
If NULL, default settings are used.}

\item{theme_mp}{The theme for the main panel. Function \code{theme} from \code{ggplot2} package.}

\item{theme_hl}{The theme for the left panel. Function \code{theme} from \code{ggplot2} package.}

\item{theme_hb}{The theme for the bottom panel. Function \code{theme} from \code{ggplot2} package.}
}
\description{
\code{plot_me} combines a heatmap of the marginal effects of a variable with the elements representing
its joint distribution with another independent variable. The heatmaps is overlaid with markers whose size
represents the number of observations. On the sides, we append histograms of the independent variables
under consideration.
}
\details{
\code{plot_me} provides a convenient way to interpret two-way interactions using
a heatmap with the histograms of the independent variables on x and y axis. Procedure allows users to customize the
plot using functions and layers from the \code{ggplot2} package. For instance, the look
of contour plot and histograms is fully customizable, but the overall style (histograms on the
left and on the bottom with contourplot in the center) is preserved for all plots.
}
\examples{
##Poisson regression with 2 variables and interaction between them
data <- data.frame(y = rpois(10000, 10), x2 = rpois(10000, 5), x1 = rpois(10000, 3))
y <- glm(y ~ x1 + x2 + x1*x2, data = data, family = "poisson")
plot_me(model = y, data = data, x = "x1", over = "x2")
}
\references{
Moral, Mert, Evgeny Sedashov, and Andrei Zhirnov. (2017) ``Taking Distributions Seriously
Interpreting the Effects of Constitutive Variables in Nonlinear Models with Interactions.'' Working paper.
}
\author{
\code{plot_me} visualizes ME procedure described in Moral, Sedashov, and Zhirnov (2017)
using the tools from \code{ggplot2} package.
}
